player/api/requirements.txt
fastapi
uvicorn
python-mpd2
psutil
